+++ src/file.hpp
@@ -14,7 +14,7 @@ struct Entry { // Big endian for cross platofr
     std::chrono::time_point<std::chrono::system_clock> ts;
     std::array<char, 6> shaddr;
     std::array<char, 6> thaddr;
-    std::uint32_t eth_proto;
+    std::uint16_t eth_proto;
 
     // IPv4 || IPv6
     std::array<char, 16> saddr;
